+++ database/migrations/2026_02_09_052940_create_event_participants_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->uuid('head_of_family_id');
             $table->foreign('head_of_family_id')->references('id')->on('head_of_families');
             $table->integer('quantity');
-            $table->decimal('total_price',10,2);
+            $table->bigInteger('total_price');
             $table->string('payment_status')->default('unpaid');
             $table->softDeletes();
             $table->timestamps();
